refactor: split appstate into crash, logging, termination helpers
